requirements.txt
# Core web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Agent orchestration
langgraph==0.0.26
langchain==0.1.4
langchain-openai==0.0.5

# Database
aiomysql==0.2.0
mysql-connector-python==8.3.0

# Redis
redis==5.0.1
aioredis==2.0.1

# Azure DevOps SDK
azure-devops==7.1.0b4

# Code parsing
tree-sitter==0.20.4

# OpenAI
openai==1.10.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
hypothesis==6.98.3
fakeredis==2.21.0

# Utilities
aiofiles==23.2.1
httpx==0.26.0
pyyaml==6.0.1
